simple_requirements.txt
# Core dependencies for basic bot functionality
aiogram==3.15.0
aiohttp==3.10.11
python-dotenv==1.0.1
requests==2.31.0

# Video processing (basic)
yt-dlp==2024.12.13
ffmpeg-python==0.2.0

# Google services (basic)
google-api-python-client==2.115.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Basic image processing
pillow>=10.0.0